+++ web/src/components/base/AuthContext.js
@@ -3,14 +3,20 @@ import React, { createContext, useContext, useState } from 'react';
 const AuthContext = createContext();
 
 export const AuthProvider = ({ children }) => {
-    const [credentials, setCredentials] = useState({});
+    const [authState, setAuthState] = useState({
+        credentials: {},
+        userEmail: '', // Explicitly include userEmail
+    });
 
     const setAuthCredentials = (email, password) => {
-        setCredentials({ email, password });
+        setAuthState({
+            credentials: { email, password },
+            userEmail: email, // Update userEmail here
+        });
     };
 
     return (
-        <AuthContext.Provider value={{ credentials, setAuthCredentials }}>
+        <AuthContext.Provider value={{ ...authState, setAuthCredentials }}>
             {children}
         </AuthContext.Provider>
     );
